stat refactor

+++ src/party/stat_calculation.h
@@ -9,7 +9,7 @@
 #include <string>
 
 void calculateActorStats(SrPartyData& srPartyMember, const CharacterRecord& charRecord, const StatBoostModifiers& statModifiers);
-u8 getGameBaseStat(std::string statName, const CharacterRecord& record);
+u32 getGameBaseStat(std::string statName, const CharacterRecord& record);
 void addStatBoosts(StatBoostModifiers& acc, const EquipmentStatBoosts& boosts);
